Extract set seanson function
